requirements.txt
Django==4.2.1
djangorestframework==3.14.0
channels==3.0.5
django-eventstream==4.5.1
django-rq==2.8.1
